Divided client's game.ts in multiple files for better scalability before doing 3,4,5,6 player game Canvas
