+++ shared/src/models/Ball.ts
@@ -185,20 +185,5 @@ export class Ball
         this.ballStart(false);
     }
 
-    public render(ctx: CanvasRenderingContext2D, color: string): void
-    {
-        const centerX = this.positionX + this.size / 2;
-        const centerY = this.positionY + this.size / 2;
-        
-        ctx.fillStyle = color;
-        ctx.save();
-        ctx.translate(centerX, centerY);
-        ctx.rotate(this.rotation);
-        ctx.translate(-this.size / 2, -this.size / 2);
-        ctx.fillRect(3, 0, 6, 3);
-        ctx.fillRect(0, 3, 12, 3);
-        ctx.fillRect(0, 6, 12, 3);
-        ctx.fillRect(3, 9, 6, 3);
-        ctx.restore();
-    }
+    
 }
